// workspaces/built-ins/src/processor.audioEncoder/types.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export type paths = Record<string, never>;
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        /**
         * @description Audio channel configuration
         * @enum {string}
         */
        channelLayout: "mono" | "stereo" | "surround" | "4.0" | "5.0" | "5.1" | "7.1" | "5.1.4" | "7.1.4";
        /**
         * @description Audio sample rate in Hz
         * @enum {number}
         */
        sampleRate: 8000 | 11025 | 12000 | 16000 | 22050 | 24000 | 32000 | 44100 | 48000 | 64000 | 88200 | 96000;
        /**
         * @description AAC encoding profile
         * @enum {string}
         */
        aacProfile: "main" | "lc" | "high";
        aac: {
            /**
             * @description discriminator enum property added by openapi-typescript
             * @enum {string}
             */
            kind: "aac";
            profile: components["schemas"]["aacProfile"];
            sampleRate: components["schemas"]["sampleRate"];
        };
        opus: {
            /**
             * @description discriminator enum property added by openapi-typescript
             * @enum {string}
             */
            kind: "opus";
        };
        codec: components["schemas"]["aac"] | components["schemas"]["opus"];
        audioEncoderConfig: {
            /** @description Unique identifier for the component */
            id: string;
            /** @description Human-readable name for the component */
            displayName: string;
            /** @description Optional notes about the component configuration */
            notes?: string;
            /** @description Name of this rendition of encoded audio */
            renditionName: string;
            channelLayout: components["schemas"]["channelLayout"];
            /** @description Bitrate in bits per second */
            bitrate: number;
            codec: components["schemas"]["codec"];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export type operations = Record<string, never>;